+++ variant_test.go
@@ -11,7 +11,7 @@ import (
 
 func variantTests(t *testing.T, variant Variant) {
 	// Check that the ID is assigned to the returned variant
-	expectedInt := 1
+	expectedInt := int64(1)
 	if variant.ID != expectedInt {
 		t.Errorf("Variant.ID returned %+v, expected %+v", variant.ID, expectedInt)
 	}
